refactor(core): single ai brief function on MR template
refs. metriport/metriport-internal#2510
